+++ gateway/src/apicast/policy/echo.lua
@@ -1,9 +1,36 @@
-local policy = require('apicast.policy')
-local _M = policy.new('Echo Policy')
+--- Echo policy
+-- Print the request back to the client and optionally set a status code.
+-- Also can interrupt the execution and skip the current phase or
+-- the whole processing of the request.
 
-function _M.rewrite()
+local _M = require('apicast.policy').new('Echo Policy')
+
+local tonumber = tonumber
+local new = _M.new
+
+function _M.new(configuration)
+  local policy = new(configuration)
+
+  if configuration then
+    policy.status = tonumber(configuration.status)
+    policy.exit = configuration.exit
+  end
+
+  return policy
+end
+
+function _M:rewrite()
+  if self.status then
+    ngx.status = self.status
+  end
+
   ngx.say(ngx.var.request)
-  ngx.exit(0)
+
+  if self.exit == 'request' then
+    return ngx.exit(ngx.status)
+  elseif self.exit == 'phase' then
+    return ngx.exit(0)
+  end
 end
 
 return _M